+++ src/main/java/de/maxhenkel/reap/Main.java
@@ -1,8 +1,8 @@
 package de.maxhenkel.reap;
 
+import de.maxhenkel.corelib.CommonRegistry;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -13,10 +13,12 @@ public class Main {
 
     public static final String MODID = "reap";
 
+    public static ServerConfig SERVER_CONFIG;
+
     public Main() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::commonSetup);
 
-        ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, Config.SERVER_SPEC);
+        SERVER_CONFIG = CommonRegistry.registerConfig(ModConfig.Type.SERVER, ServerConfig.class, true);
     }
 
     @SubscribeEvent