+++ changelog.md
@@ -1,1 +1,1 @@
-- Internal changes
+- Updated to 1.16.2